Added authentication and authorization + minor fixes

+++ src/handles/event.rs
@@ -1,11 +1,12 @@
 use actix_web::{HttpResponse, delete, get, patch, post, web::{Data, Json, Path, Query}};
+use actix_web_httpauth::middleware::HttpAuthentication;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, StringWithSeparator, formats::CommaSeparator};
 use serde_aux::field_attributes::deserialize_default_from_empty_object;
 use utoipa_actix_web::{scope, service_config::ServiceConfig};
 
 use crate::{
-        domain::{event::{model::*, repository::EventRepository}, utils::Offset}, infrastructure::{db::event::PgEventRepository, provider::PgProvider}
+        domain::{event::{model::*, repository::EventRepository}, user::model::UserRole, utils::Offset}, handles::auth::{AuthenticatedUser, validator}, infrastructure::{db::event::PgEventRepository, provider::PgProvider}
 };
 
 use super::error::Result;
@@ -13,12 +14,15 @@ use utoipa::ToSchema;
 
 pub fn event_app_config(cfg: &mut ServiceConfig) {
         cfg
-        .service(scope::scope("/events")
-        .service(get_event)
-        .service(create_event)
-        .service(update_event_status)
-        .service(list_events)
-        .service(delete_event)
+                .service(scope::scope("/events")
+                .service(get_event)
+                .service(list_events)
+                .service(scope::scope("")
+                        .wrap(HttpAuthentication::bearer(validator))
+                        .service(create_event)
+                        .service(update_event_status)
+                        .service(delete_event)
+                )
         );
 }
 
@@ -47,7 +51,11 @@ async fn get_event(provider: Data<PgProvider>, id: Path<EventId>) -> Result<Http
 
 #[utoipa::path]
 #[post("")]
-async fn create_event(provider: Data<PgProvider>, event: Json<NewEvent>) -> Result<HttpResponse> {
+async fn create_event(provider: Data<PgProvider>, auth_user: AuthenticatedUser, event: Json<NewEvent>) -> Result<HttpResponse> {
+        if auth_user.role == UserRole::User || auth_user.id != event.organizer_id {
+                return Ok(HttpResponse::Forbidden().finish());
+        }
+
         let repo = provider.provide_event_repository::<PgEventRepository>();
 
         let event =
@@ -73,8 +81,12 @@ async fn create_event(provider: Data<PgProvider>, event: Json<NewEvent>) -> Resu
 
 #[utoipa::path(params(("event_id" = EventId, Path)))]
 #[delete("/{event_id}")]
-async fn delete_event(provider: Data<PgProvider>, id: Path<EventId>) -> Result<HttpResponse> {
+async fn delete_event(provider: Data<PgProvider>, auth_user: AuthenticatedUser, id: Path<EventId>) -> Result<HttpResponse> {
         let repo = provider.provide_event_repository::<PgEventRepository>();
+        if auth_user.role == UserRole::User || (auth_user.role == UserRole::Organizer && repo.get(*id).await?.is_none_or(|v| v.organizer_id == *id)) {
+                return Ok(HttpResponse::Forbidden().finish());
+        }
+
         let event = repo.delete(*id).await?;
 
         let res = match event {
@@ -96,7 +108,11 @@ struct EventStatusBody {
 
 #[utoipa::path(params(("event_id" = EventId, Path)))]
 #[patch("/{event_id}")]
-async fn update_event_status(provider: Data<PgProvider>, id: Path<EventId>, status: Json<EventStatusBody>) -> Result<HttpResponse> {
+async fn update_event_status(provider: Data<PgProvider>, auth_user: AuthenticatedUser, id: Path<EventId>, status: Json<EventStatusBody>) -> Result<HttpResponse> {
+        if auth_user.role < UserRole::Admin {
+                return Ok(HttpResponse::Forbidden().finish());
+        }
+
         let repo = provider.provide_event_repository::<PgEventRepository>();
         let event = repo.update(*id, EventUpdate::Status(status.into_inner().status)).await?;
