+++ src/handles/user.rs
@@ -1,21 +1,21 @@
 use actix_web::{HttpResponse, delete, get, patch, post, web::{Data, Json, Path, Query}};
+use actix_web_httpauth::middleware::HttpAuthentication;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, StringWithSeparator, formats::CommaSeparator};
 use serde_aux::field_attributes::deserialize_default_from_empty_object;
 use utoipa_actix_web::{scope, service_config::ServiceConfig};
 
 use crate::{
-        domain::{user::{model::*, repository::UserRepository}, utils::Offset}, infrastructure::{db::user::PgUserRepository, provider::PgProvider}
+        domain::{user::{model::*, repository::UserRepository}, utils::Offset}, handles::auth::{AuthenticatedUser, Credentials, validator}, infrastructure::{db::user::PgUserRepository, provider::PgProvider}
 };
 
 use super::error::Result;
 use utoipa::ToSchema;
 use argon2::{
-    password_hash::{
-        rand_core::OsRng,
-        PasswordHasher, SaltString
+    Argon2, PasswordVerifier, password_hash::{
+        PasswordHasher, SaltString, rand_core::OsRng
     },
-    Argon2
+    PasswordHash
 };
 
 pub fn user_app_config(cfg: &mut ServiceConfig) {
@@ -23,10 +23,14 @@ pub fn user_app_config(cfg: &mut ServiceConfig) {
         .service(scope::scope("/users")
                 .service(get_user)
                 .service(create_user)
-                .service(delete_user)
-                .service(update_user_role)
                 .service(list_users)
-                .configure(super::favorite::favorite_app_config)
+                .service(login)
+                .service(scope::scope("")
+                        .wrap(HttpAuthentication::bearer(validator))
+                        .service(update_user_role)
+                        .service(delete_user)
+                        .configure(super::favorite::favorite_app_config)
+                )
         );
 }
 
@@ -89,7 +93,11 @@ async fn create_user(provider: Data<PgProvider>, mut credentials: Json<UserCrede
 
 #[utoipa::path(params(("user_id" = UserId, Path)))]
 #[delete("/{user_id}")]
-async fn delete_user(provider: Data<PgProvider>, id: Path<UserId>) -> Result<HttpResponse> {
+async fn delete_user(provider: Data<PgProvider>, auth_user: AuthenticatedUser, id: Path<UserId>) -> Result<HttpResponse> {
+        if auth_user.role < UserRole::Admin && auth_user.id != *id {
+                return Ok(HttpResponse::Forbidden().finish());
+        }
+
         let repo = provider.provide_user_repository::<PgUserRepository>();
         let user = repo.delete(*id).await?;
 
@@ -112,7 +120,11 @@ struct UserRoleBody {
 
 #[utoipa::path(params(("user_id" = UserId, Path)))]
 #[patch("/{user_id}")]
-async fn update_user_role(provider: Data<PgProvider>, id: Path<UserId>, role: Json<UserRoleBody>) -> Result<HttpResponse> {
+async fn update_user_role(provider: Data<PgProvider>, auth_user: AuthenticatedUser , id: Path<UserId>, role: Json<UserRoleBody>) -> Result<HttpResponse> {
+        if auth_user.role < UserRole::Admin {
+                return Ok(HttpResponse::Forbidden().finish());
+        }
+
         let repo = provider.provide_user_repository::<PgUserRepository>();
         let user = repo.update(*id, UserUpdate::Role(role.into_inner().role)).await?;
 
@@ -158,4 +170,32 @@ async fn list_users(provider: Data<PgProvider>, query: Query<ListUsersQuery>) ->
         let users = repo.list(query.offset.clone(), &query.filter, &query.order_by).await?;
 
         Ok(HttpResponse::Ok().json(ListUsersResponse {users}))
+}
+
+#[derive(Debug, Serialize, utoipa::ToResponse)]
+struct JwtResponse {
+        jwt: String
+}
+
+#[utoipa::path]
+#[post("/login")]
+async fn login(provider: Data<PgProvider>, credentials: Json<UserCredentials>) -> Result<HttpResponse> {
+        let repo = provider.provide_user_repository::<PgUserRepository>();
+        let credentials = credentials.into_inner();
+        let user = repo.login(credentials.login).await?;
+
+        let res = match user {
+                Some(user) => {
+                        if Argon2::default().verify_password(credentials.password.as_bytes(), &PasswordHash::new(&user.password_hash)?).is_err() {
+                                return Ok(HttpResponse::Forbidden().finish())
+                        }
+                        let jwt = Credentials::new(user.into()).encode()?;
+                        HttpResponse::Ok().json(JwtResponse {jwt})
+                }
+                None => {
+                        HttpResponse::NotFound().finish()
+                }
+        };
+
+        Ok(res)
 }
